Incorporate new syntax using F" and simplify
We have suggested switching to F" ... " as the indicator for an interpolated string literal, and eliminated the multi-line string literal option for now.
One option worth considering in our view is to use {" ... "} as the indicator for an interpolated string literal, thereby avoiding the somewhat odd notion of a 2-character delimiter where one of the characters is a "regular" letter (i.e. 'F').